requirements.txt
requests
pyyaml
lxml
coverage
flake8
bandit
prometheus_client
